+++ src/AssemblyLoadContext/AssemblyLoadContextInitializer.cs
@@ -23,6 +23,9 @@
         {
             Dictionary<string, Version> sharedAssemblies = new()
             {
+                { "Microsoft.Extensions.Primitives", new Version("7.0.0.0") },
+                { "Microsoft.Graph.Beta", new Version("4.68.0.0") },
+                { "Microsoft.Graph.Core", new Version("2.0.14.0") },
                 { "Microsoft.Identity.Client", new Version("4.48.0.0") },
                 { "Microsoft.Identity.Client.Extensions.Msal", new Version("2.25.0.0") },
                 { "Microsoft.IdentityModel.Abstractions", new Version("6.22.0.0") }