+++ src/Authentication/Factories/IClientFactory.cs
@@ -1,6 +1,7 @@
 ﻿namespace AutoBrew.PowerShell.Factories
 {
     using Microsoft.Graph;
+    using Microsoft.Store.PartnerCenter;
     using Models.Authentication;
 
     /// <summary>
@@ -9,10 +10,18 @@
     public interface IClientFactory
     {
         /// <summary>
-        /// Creates a new instance of the <see cref="GraphServiceClient" /> used to communicate with Microsoft Graph.
+        /// Creates a new instance of the <see cref="GraphServiceClient" /> class used to communicate with Microsoft Graph.
         /// </summary>
         /// <param name="account">An instance of the <see cref="ModuleAccount" /> class that provides information used to authenticate.</param>
-        /// <returns>An instance of the <see cref="GraphServiceClient" /> used to communicate with Microsoft Graph.</returns>
+        /// <returns>An instance of the <see cref="GraphServiceClient" /> class used to communicate with Microsoft Graph.</returns>
         GraphServiceClient CreateGraphServiceClient(ModuleAccount account);
+
+        /// <summary>
+        /// Creates a new instance of the <see cref="PartnerOperations" /> class used to communicate with Partner Center.
+        /// </summary>
+        /// <param name="account">An instance of the <see cref="ModuleAccount" /> class that provides information used to authenticate.</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>An instance of the <see cref="PartnerOperations" /> class used to communicate with partner Center.</returns>
+        Task<IPartner> CreatePartnerOperationsAsync(ModuleAccount account, CancellationToken cancellationToken = default);
     }
 }